requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
black==24.4.2
cffi==1.16.0
click==8.1.7
flake8==7.0.0
greenlet==3.0.3
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.2
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pycparser==2.22
pyflakes==3.2.0
SQLAlchemy==2.0.30
SQLAlchemy-Utils==0.41.2
tomli==2.0.1
typing_extensions==4.11.0
